[#124] refactor: user 관련 API 분리
유저 목록을 주는 API와 로그인 API를 분리합니다.
유저 추가 API는 사용하지 않을 것 같아 제거합니다.
