+++ BE/src/main/java/kr/codesquad/todo9/api/TodoAPIController.java
@@ -3,23 +3,15 @@ package kr.codesquad.todo9.api;
 import kr.codesquad.todo9.domain.*;
 import kr.codesquad.todo9.error.exception.BoardNotFoundException;
 import kr.codesquad.todo9.error.exception.LogNotFoundException;
-import kr.codesquad.todo9.error.exception.UserNotFoundException;
 import kr.codesquad.todo9.repository.BoardRepository;
-import kr.codesquad.todo9.repository.UserRepository;
 import kr.codesquad.todo9.requestobject.ContentsObject;
 import kr.codesquad.todo9.requestobject.MoveCardObject;
-import kr.codesquad.todo9.responseobject.Result;
-import kr.codesquad.todo9.utils.JwtUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
-import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 import java.util.ArrayList;
 import java.util.Collections;
@@ -30,45 +22,11 @@ public class TodoAPIController {
 
     private static final Logger log = LoggerFactory.getLogger(TodoAPIController.class);
     private static final Long defaultBoardId = 1L;
-    private static final Long defaultUserId = 1L;
 
-    private final UserRepository userRepository;
     private final BoardRepository boardRepository;
 
-    public TodoAPIController(UserRepository userRepository, BoardRepository boardRepository) {
-        this.userRepository = userRepository;
+    public TodoAPIController(BoardRepository boardRepository) {
         this.boardRepository = boardRepository;
-    }
-
-    @PostMapping("/login")
-    public ResponseEntity<Result> login(HttpServletResponse httpServletResponse) {
-        Result result = new Result(true, "success");
-        User user = userRepository.findById(defaultUserId).orElseThrow(UserNotFoundException::new);
-        String jws = JwtUtils.createJws(user);
-        log.debug("user: {}", user);
-        log.debug("jws: {}", jws);
-
-        Cookie jwsCookie = new Cookie("jws", jws);
-        httpServletResponse.addCookie(jwsCookie);
-
-        return new ResponseEntity<>(result, HttpStatus.OK);
-    }
-
-    @GetMapping("/user/list")
-    public Iterable<User> showUserList() {
-        return userRepository.findAll();
-    }
-
-    @PostMapping("/user/{name}")
-    public Result addUser(@PathVariable String name) {
-        User newUser = new User();
-        newUser.setUsername(name);
-        log.debug("newUser: {}", newUser);
-
-        newUser = userRepository.save(newUser);
-        log.debug("newUser: {}", newUser);
-
-        return new Result(true, "success");
     }
 
     @GetMapping("/board/{boardId}/column/{boardKey}/card/list")
